Fully decoupled codegen from clang:
  * Added forward declarations of clang types in ir/types.hpp

  * Extracted module_info_t, cls_info_t, and fnt_info_t out of clair-c2py/wdata.hpp
