+++ src/tools/ir/types.cpp
@@ -7,6 +7,10 @@
 
 #include <clang/AST/DeclTemplate.h>
 #include <clang/Lex/Lexer.h>
+#include <clang/AST/ASTContext.h>
+#include <clang/AST/Decl.h>
+#include <clang/AST/DeclCXX.h>
+#include <clang/AST/Type.h>
 #include <llvm/ADT/SmallVector.h>
 #include <llvm/ADT/StringMap.h>
 #include <fmt/format.h>
